+++ src/kernel/bpf/mod.rs
@@ -1,0 +1,10 @@
+mod btf;
+mod prog;
+mod sys;
+
+use btf::Btf;
+
+pub(super) use prog::BpfInfoCache;
+pub(super) use prog::BpfProg;
+#[cfg(test)]
+pub(super) use prog::BpfTag;